refactor: move parseOembedUrl to the server
